requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
google-cloud-storage==2.14.0
python-multipart==0.0.6